requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
contextlib2==21.6.0
dj-database-url==2.1.0
Django==4.0.4
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
dj-rest-auth==5.0.2
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
schema==0.7.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
# urllib3==2.1.0
whitenoise==6.6.0
urllib3==1.26.6
